fix confirm delete

+++ resources/views/permission/index.blade.php
@@ -173,9 +173,10 @@
         $(document).on('click', '.delete', function() {
             var id = $(this).data('id')
             console.log(id);
-            confirm('Are you sure you want to delete this permission?');
+            var result = confirm('Are you sure you want to delete this permission?');
 
-            $.ajax({
+            if(result) {
+                $.ajax({
                 url: '{{ route("permission.destroy") }}',
                 method: "DELETE",
                 data: {
@@ -185,6 +186,7 @@
                     table.DataTable().ajax.reload();
                 }
             })
+            }
         })
 
         $(document).on('click', '.edit', function(e) {
