+++ resources/views/user/index.blade.php
@@ -198,9 +198,10 @@
         $(document).on('click', '.delete', function() {
             var id = $(this).data('id')
             console.log(id);
-            confirm('Are you sure you want to delete this user?');
+            var result = confirm('Are you sure you want to delete this user?');
 
-            $.ajax({
+            if(result) {
+                $.ajax({
                 url: '{{ route("user.destroy") }}',
                 method: "DELETE",
                 data: {
@@ -210,6 +211,7 @@
                     table.DataTable().ajax.reload();
                 }
             })
+            }
         })
 
         $(document).on('click', '.edit', function(e) {
